exit fix

+++ srcs/libft/ft_atoi.c
@@ -6,13 +6,14 @@
 /*   By: mbouzaie <mbouzaie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/10 14:09:57 by mbouzaie          #+#    #+#             */
-/*   Updated: 2020/05/12 18:51:08 by mbouzaie         ###   ########.fr       */
+/*   Updated: 2021/03/13 16:21:30 by mbouzaie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/libft.h"
+#include <stdio.h>
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
 	int					sign;
 	unsigned long int	result;
@@ -29,12 +30,12 @@ int	ft_atoi(const char *str)
 	{
 		if ((result > 922337203685477580 || (result == 922337203685477580
 			&& (*str - '0') > 7)) && sign == 1)
-			return (-1);
+				return (-1);
 		else if ((result > 922337203685477580 || (result == 922337203685477580
 			&& (*str - '0') > 8)) && sign == -1)
 			return (0);
 		result = (result * 10) + (*str - '0');
 		str++;
 	}
-	return ((int)(result * sign));
+	return (result * sign);
 }
