+++ srcs/libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: mbouzaie <mbouzaie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/03 14:55:29 by mbouzaie          #+#    #+#             */
-/*   Updated: 2021/02/19 14:28:37 by mbouzaie         ###   ########.fr       */
+/*   Updated: 2021/03/13 16:20:52 by mbouzaie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,7 +54,7 @@ char				*ft_strrchr(const char *s, int c);
 char				*ft_strnstr(const char *haystack, const char *needle, \
 								size_t len);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
-int					ft_atoi(const char *str);
+long int			ft_atoi(const char *str);
 int					ft_isupper(int c);
 int					ft_islower(int c);
 int					ft_isalpha(int c);
